+++ src/Filters/ColumnIntersectsCollection.php
@@ -37,7 +37,7 @@ class ColumnIntersectsCollection extends Equals
         $this->columnName = $columnName;
         $this->collection = $collectionToCheckForIntersection;
 
-        parent::__construct($columnName, "@{".$collectionToCheckForIntersection->getUniqueReference().$columnName."}");
+        parent::__construct($columnName, "@{".$collectionToCheckForIntersection->getUniqueReference()."_".$columnName."}");
     }
 
     /**
@@ -53,6 +53,6 @@ class ColumnIntersectsCollection extends Equals
         // While this behaviour could (and perhaps more efficiently) be achieved using `intersectWith` we instead
         // do a `joinWith` and pull the target column up so that we can filter in the normal way.
         // This allows this filter to be combined with the 'not' filter and behave as expected.
-        $collection->joinWith($this->collection, $this->columnName, $this->columnName, [$this->columnName => $this->collection->getUniqueReference().$this->columnName]);
+        $collection->joinWith($this->collection, $this->columnName, $this->columnName, [$this->columnName => $this->collection->getUniqueReference()."_".$this->columnName]);
     }
 }