+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.2.x
 
+* Change:   Changed MySqlLongStringColumn to store as an SQL 'LONGTEXT' column rather than a shorter 'TEXT' column
 
 ### 1.2.2